seperated conn management code from tcpx_fabric.c
