+++ prov/tcp/Makefile.include
@@ -2,6 +2,7 @@ if HAVE_TCP
 _tcp_files = \
 	prov/tcp/src/tcpx_fabric.c	\
 	prov/tcp/src/tcpx_attr.c	\
+	prov/tcp/src/tcpx_conn_mgr.c	\
 	prov/tcp/src/tcpx_domain.c	\
 	prov/tcp/src/tcpx_ep.c		\
 	prov/tcp/src/tcpx_cq.c		\
